TodoDetail→TodoModalへ名称変更と中身修正(CRUDのモードによって出し分ける)
